fix update logic
