+++ examples/complete-with-update/main.tf
@@ -1,0 +1,12 @@
+resource "shell_script" "update_test" {
+  lifecycle_commands {
+    create = file("${path.module}/scripts/create.sh")
+    read   = file("${path.module}/scripts/read.sh")
+    update = file("${path.module}/scripts/update.sh")
+    delete = file("${path.module}/scripts/delete.sh")
+  }
+
+  environment = {
+    DESCRIPTION = "somedescription"
+  }
+}
