+++ examples/complete-with-update/scripts/create.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+id=$RANDOM
+/bin/cat <<END >$id.json
+  {"id": "$id", "description": "$DESCRIPTION"}
+END
+cat $id.json
