+++ examples/complete-with-update/scripts/delete.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+IN=$(cat)
+id=$(echo $IN | jq -r .id)
+rm ${id}.json
