+++ examples/complete-with-update/scripts/read.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+IN=$(cat)
+id=$(echo $IN | jq -r .id)
+cat ${id}.json
