+++ examples/complete-with-update/scripts/update.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+IN=$(cat)
+id=$(echo $IN | jq -r .id)
+/bin/cat <<END >$id.json
+  {"id": "$id", "description": "$DESCRIPTION"}
+END
